task: Created static method and test case for Median. Added CSV Data

+++ Calculator/Square.py
@@ -1,2 +1,2 @@
 def squaring(a):
-    return int(a) ** 2
+    return float(a) ** 2
